pyproject.toml
[tool.poetry]
name = "alpaca"
version = "0.1.0"
description = "Discord bot for pixel games"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
discord-py = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
